Remove unnecessary common/utils.js file
